requirements.txt
pip==24.1.2
python-dotenv==0.19.0
python-telegram-bot==13.7